recompile 2

+++ Mlib/Stats/Histogram.hpp
@@ -41,7 +41,7 @@ public:
                 return hist_.length() - 1;
             }
         }
-        return ((hist_.length() - 1) * (data - mi_)) / (ma_ - mi_);
+        return (size_t)(((hist_.length() - 1) * (data - mi_)) / (ma_ - mi_));
     }
 private:
     Array<size_t> hist_;
